refactor : VirtualKeyboard를 사이트의 용도에 맞게 코드 수정
