+++ src/components/VirtualKeyboard/VKBoxHeader/VKBoxHeader.helper.ts
