+++ src/components/VirtualKeyboard/VkBox/index.ts
@@ -1,0 +1,2 @@
+export * from "./VKBox";
+export { default } from "./VKBox";
